chore: move base test case to common app
